requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74
pillow==10.0.0
numpy==1.25.2
pdf2image==1.16.3

